feat: Kafka consumer app
Need an app that can demonstrate the messages that are arriving
in each region, so this will just print messages to stdout.

The output includes the topic name, partition and offset info, and
the timestamp - this will make it easier to understand where the
flow of messages are coming from.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>